Fixing the code, something is OFF here
Need to dig into more, I think the test data needs help. I may have made it worse ... but the model was diverging initially so I upped the sites and plots, but the code was not written to automatically update (10 was in lots of places ... where it should have been J or S... not sure).